+++ mayan/apps/converter/static/converter/js/mayan_image.js
@@ -3,12 +3,10 @@
 class MayanImage {
     constructor (options) {
         this.element = options.element;
-        this.load();
     }
 
     static async setup (options) {
         this.options = options || {};
-        this.options.templateInvalidImage = this.options.templateInvalidImage || '<span>Error loading image</span>';
 
         $().fancybox({
             afterShow: function (instance, current) {
@@ -29,31 +27,34 @@ class MayanImage {
     }
 
     static async initialize () {
-        $('img.lazy-load').lazyload({
-            appear: async function(elements_left, settings) {
-                new MayanImage({element: $(this)});
-            },
-            threshold: 400,
+        $('img.lazy-load').each(async function(index) {
+            const $this = $(this);
+
+            $this.attr('loading', 'lazy');
+            const dataURL = $this.attr('data-url');
+            const url = new URL(dataURL, window.location.origin);
+            $this.attr('src', url);
         });
 
-        $('img.lazy-load-carousel').lazyload({
-            appear: async function(elements_left, settings) {
-                new MayanImage({element: $(this)});
-            },
-            container: $('#carousel-container'),
-            threshold: 2000,
+        $('img.lazy-load-carousel').each(async function(index) {
+            const $this = $(this);
+
+            $this.attr('loading', 'lazy');
+            const dataURL = $this.attr('data-url');
+            const url = new URL(dataURL, window.location.origin);
+            $this.attr('src', url);
         });
 
-        $('.lazy-load').on('load', function() {
+        $('.lazy-load').on('load', async function() {
             const $this = $(this);
 
             $this.siblings('.lazyload-spinner-container').remove();
             $this.removeClass('lazy-load pull-left');
             clearTimeout(MayanImage.timer);
-            MayanImage.timer = setTimeout(MayanImage.timerFunction, 250);
+            MayanImage.timer = setTimeout(MayanImage.timerFunction, 50);
         });
 
-        $('.lazy-load-carousel').on('load', function() {
+        $('.lazy-load-carousel').on('load', async function() {
             const $this = $(this);
 
             $this.siblings('.lazyload-spinner-container').remove();
@@ -64,33 +65,6 @@ class MayanImage {
     static timerFunction () {
         $.fn.matchHeight._update();
     }
-
-    async load () {
-        const self = this;
-        const container = this.element.parent().parent().parent();
-        const dataURL = this.element.attr('data-url');
-
-        if (dataURL === '') {
-            container.html(MayanImage.options.templateInvalidImage);
-        } else {
-            this.element.attr('src', dataURL);
-            setTimeout(function () {
-                self.element.on('error', function () {
-                    // Check the .complete property to see if it is a real
-                    // error or it was a cached image
-                    if (this.complete === false) {
-                        // It is a cached image, set the src attribute to
-                        // trigger its display.
-                        this.src = dataURL;
-                    } else {
-                        container.html(
-                            MayanImage.options.templateInvalidImage
-                        );
-                    }
-                });
-            }, 1);
-        }
-    };
 }
 
 MayanImage.timer = setTimeout(null);