+++ HISTORY.rst
@@ -92,6 +92,8 @@
 - Add notification detail API view.
 - Unify condition field model mixin and helper methods.
 - Remove URI.js. Instead use the JavaScript native URL class.
+- Remove jQuery Lazy Load libary ``jquery-lazyload``.
+- Remove jquery-scrollstop.
 
 4.6.2 (2024-03-04)
 ==================